+++ common-lib/src/main/java/com/bank/common/dto/contracts/exchange/ExchangeRateDTO.java
@@ -2,12 +2,14 @@ package com.bank.common.dto.contracts.exchange;
 
 import lombok.*;
 
+import java.math.BigDecimal;
+
 @Data
 @Builder
 @NoArgsConstructor
 @AllArgsConstructor
 public class ExchangeRateDTO {
     private String currency;
-    private Double buyRate;
-    private Double sellRate;
+    private BigDecimal buyRate;
+    private BigDecimal sellRate;
 }
