+++ common-lib/src/main/java/com/bank/common/dto/contracts/transfer/TransferResponse.java
@@ -2,6 +2,8 @@ package com.bank.common.dto.contracts.transfer;
 
 import lombok.*;
 
+import java.math.BigDecimal;
+
 @Data
 @Builder
 @NoArgsConstructor
@@ -10,5 +12,5 @@ public class TransferResponse {
     private Long transferId;
     private String status;
     private String message;
-    private Double convertedAmount;
+    private BigDecimal convertedAmount;
 }
