+++ exchange-service/src/main/java/com/bank/exchange/service/ExchangeService.java
@@ -2,10 +2,11 @@ package com.bank.exchange.service;
 
 import com.bank.common.dto.contracts.exchange.ExchangeRateDTO;
 
+import java.math.BigDecimal;
 import java.util.List;
 
 public interface ExchangeService {
     List<ExchangeRateDTO> getAllRates();
-    void updateRate(String currency, Double buyRate, Double sellRate);
-    Double convert(Double amount, String fromCurrency, String toCurrency);
+    void updateRate(String currency, BigDecimal buyRate, BigDecimal sellRate);
+    BigDecimal convert(BigDecimal amount, String fromCurrency, String toCurrency);
 }
