+++ transfer-service/src/main/java/com/bank/transfer/client/ExchangeClientFallback.java
@@ -5,12 +5,14 @@ import com.bank.common.dto.contracts.exchange.ConversionRequest;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.stereotype.Component;
 
+import java.math.BigDecimal;
+
 @Slf4j
 @Component
 public class ExchangeClientFallback implements ExchangeClient {
 
     @Override
-    public ApiResponse<Double> convert(ConversionRequest request) {
+    public ApiResponse<BigDecimal> convert(ConversionRequest request) {
         log.error("Exchange service is unavailable. Cannot perform currency conversion from {} to {}",
             request.getFromCurrency(), request.getToCurrency());
         return ApiResponse.error("Exchange service is unavailable. Please try again later.");
